Update to v1core0.3.13

+++ contracts/test/strategies/balancer/TestBalancerLongStrategy.sol
@@ -47,7 +47,7 @@ contract TestBalancerLongStrategy is BalancerExternalLongStrategy {
         return getVault();
     }
 
-    function testGetPoolReserves(address) public view returns(uint128[] memory) {
+    function testGetPoolReserves(address) public view returns(uint256[] memory) {
         return getPoolReserves();
     }
